+++ taf/tools/repo/__init__.py
@@ -138,11 +138,10 @@ def clone_repo_command():
     @click.option("--path", help="Authentication repository's location. If not specified, calculated by combining repository's name specified in info.json and library dir")
     @click.option("--library-dir", default=None, help="Directory where target repositories and, optionally, authentication repository are located. If not specified, set to the current directory")
     @click.option("--from-fs", is_flag=True, default=False, help="Indicates if we want to clone a repository from the filesystem")
-    # JMC: Addition of --no-deps:
+    @click.option("--bare", is_flag=True, default=False, help="Clone repositories as bare repositories")
     @click.option("--no-deps", is_flag=True, default=False, help="Optionally disables updating of dependencies")
-    # JMC: Addition of --no-upstream
     @click.option("--upstream/--no-upstream", default=False, help="Skips comparison with remote repositories upstream")
-    def clone(path, url, library_dir, from_fs, expected_repo_type, scripts_root_dir, profile, format_output, exclude_target, strict, upstream, no_deps):
+    def clone(path, url, library_dir, from_fs, expected_repo_type, scripts_root_dir, profile, format_output, exclude_target, strict, bare, upstream, no_deps):
         if profile:
             start_profiling()
 
@@ -156,6 +155,7 @@ def clone_repo_command():
             scripts_root_dir=scripts_root_dir,
             excluded_target_globs=exclude_target,
             strict=strict,
+            bare=bare,
             no_upstream=not upstream,
             no_deps=no_deps,
         )
@@ -268,9 +268,10 @@ def validate_repo_command():
     @click.option("--no-deps", is_flag=True, default=False, help="Optionally disables updating of dependencies")
     def validate(path, library_dir, from_commit, from_latest, exclude_target, strict, no_targets, no_deps):
         auth_repo = AuthenticationRepository(path=path)
+        bare = auth_repo.is_bare_repository
         if from_latest:
             from_commit = auth_repo.last_validated_commit
-        validate_repository(path, library_dir, from_commit, exclude_target, strict, no_targets, no_deps)
+        validate_repository(path, library_dir, from_commit, exclude_target, strict, bare, no_targets, no_deps)
     return validate